Fix Radix HoverCard hydration by splitting ProjectTag client/server boundary
Radix `asChild` + Slot's cloneElement doesn't hydrate correctly when the
slotted child is created in a Server Component and passed across the RSC
boundary (regression in Next.js 16.2+). Move `ProjectTag` into its own
client file accepting a pre-computed `url` string, and keep
`ProjectTagGroup` as a Server Component that consumes `buildPageURL` and
passes the resulting URL down. Also mark `CompactTagList` as a client
component for the same reason.

This avoids passing the `buildPageURL` function closure across the RSC
boundary, which broke `/projects` after the previous blanket `"use client"`
on `project-tag.tsx`.
